# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Forge CI/CD API
  description: Self-hosted CI/CD control plane. Array responses, error envelope and in-process rate limits follow docs/contracts/API_CONTRACT.md and docs/API.md current compatibility mode.
  license:
    name: FerrPOINT Proprietary Source-Available Evaluation License v1.0
    url: https://github.com/FerrPOINT/CI-CD/blob/main/LICENSE
  version: 0.1.0
paths:
...
  /api/v1/auth/login:
    post:
      tags:
      - auth
      operationId: auth_login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: ''
  /api/v1/auth/logout:
    post:
      tags:
      - auth
      operationId: auth_logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
  /api/v1/auth/principal:
    get:
      tags:
      - auth
      operationId: auth_principal
      responses:
        '200':
          description: ''
        '401':
          description: ''
  /api/v1/auth/refresh:
    post:
      tags:
      - auth
      operationId: auth_refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: ''
  /api/v1/catalog/available-repositories:
    get:
      tags:
      - namespaces
      operationId: forge_unbound_repository_catalog
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPage'
  /api/v1/catalog/repositories/{id}:
    get:
      tags:
      - namespaces
      operationId: forge_repository_catalog_get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogRepository'
...
  /api/v1/catalog/repositories/{id}/group:
    put:
      tags:
      - namespaces
      operationId: forge_attach_repository
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachRepository'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachReadback'
        '409':
          description: Already attached or conflicting original operation
  /api/v1/catalog/repositories/{id}/pulls:
    get:
      tags:
      - pulls
      operationId: forge_repository_catalog_pulls
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PullRequest'
    post:
      tags:
      - pulls
      operationId: forge_repository_catalog_create_pull
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePull'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullRequest'
  /api/v1/catalog/repositories/{id}/pulls/{number}:
    get:
      tags:
      - pulls
      operationId: forge_repository_catalog_pull
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: number
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullRequest'
  /api/v1/catalog/repositories/{id}/pulls/{number}/action:
    post:
      tags:
      - pulls
      operationId: forge_repository_catalog_pull_action
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: number
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrAction'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullRequest'
  /api/v1/catalog/repositories/{id}/pulls/{number}/tasks:
    get:
      tags:
      - namespaces
      operationId: forge_pull_task_links
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: number
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PullTaskLink'
    post:
      tags:
      - namespaces
      operationId: forge_link_task_pull
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: number
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRef'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRef'
...
  /api/v1/git-groups/{id}/repositories:
    get:
      tags:
      - namespaces
      operationId: forge_repository_catalog_list
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPage'
    post:
      tags:
      - namespaces
      operationId: forge_repository_catalog_create
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRepository'
        required: true
      responses:
        '200':
          description: Original ID replay or created catalog repository
        '409':
          description: Payload or namespace conflict
...
  /api/v1/internal/git-push:
    post:
      tags:
      - git
      summary: |-
        Internal endpoint called by the generated `post-receive` hook.
        Finds a project whose repository_url points at this repo and triggers a pipeline.
      operationId: internal_git_push
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GitPushEvent'
        required: true
      responses:
        '200':
          description: Pipeline trigger decision
        '400':
          description: ''
        '401':
          description: ''
...
  /api/v1/jobs/{job_id}/attempts:
    get:
      tags:
      - jobs
      operationId: list_job_attempts
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobAttempt'
        '404':
          description: ''
  /api/v1/jobs/{job_id}/attempts/{attempt_id}/logs:
    get:
      tags:
      - jobs
      operationId: list_attempt_logs
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: attempt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobLog'
        '404':
          description: ''
  /api/v1/jobs/{job_id}/attempts/{attempt_id}/logs/page:
    get:
      tags:
      - jobs
      summary: Bounded page of logs for a concrete attempt. Preserves the legacy array endpoint.
      operationId: list_attempt_logs_page
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: attempt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: after
        in: query
        description: Return log rows with sequence greater than this value.
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int32
      - name: before
        in: query
        description: |-
          Tail-window cursor: return the newest rows with sequence less than
          this value (descending fetch, returned in ascending order).
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int32
      - name: limit
        in: query
        description: Page size. Default and maximum are 200 rows.
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      - name: q
        in: query
        description: Optional case-insensitive substring filter for message text.
        required: false
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobLogPage'
        '400':
          description: ''
        '404':
          description: ''
  /api/v1/jobs/{job_id}/logs:
    get:
      tags:
      - jobs
      operationId: list_logs
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobLog'
    post:
      tags:
      - jobs
      operationId: append_log
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppendLog'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobLog'
        '413':
          description: log append body exceeds 1 MiB
  /api/v1/jobs/{job_id}/logs/page:
    get:
      tags:
      - jobs
      summary: Bounded page of logs for the active or latest attempt.
      operationId: list_logs_page
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: after
        in: query
        description: Return log rows with sequence greater than this value.
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int32
      - name: before
        in: query
        description: |-
          Tail-window cursor: return the newest rows with sequence less than
          this value (descending fetch, returned in ascending order).
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int32
      - name: limit
        in: query
        description: Page size. Default and maximum are 200 rows.
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      - name: q
        in: query
        description: Optional case-insensitive substring filter for message text.
        required: false
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobLogPage'
        '400':
          description: ''
        '404':
          description: ''
  /api/v1/jobs/{job_id}/logs/stream:
    get:
      tags:
      - jobs
      summary: 'SSE live log stream: emits existing lines, then polls for new ones.'
      operationId: job_log_stream
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: after
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: text/event-stream of job log lines
  /api/v1/jobs/{job_id}/retry:
    post:
      tags:
      - jobs
      operationId: retry_job
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          description: ''
  /api/v1/jobs/{job_id}/start:
    post:
      tags:
      - jobs
      summary: 'Starts a manual (`when: manual`) job — approval gate (GitLab parity).'
      operationId: start_manual_job
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualJobStartResult'
        '404':
          description: ''
        '409':
          description: job is not a waiting manual job
  /api/v1/jobs/{job_id}/status:
    post:
      tags:
      - jobs
      operationId: change_job_status
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeStatus'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          description: ''
...
  /api/v1/namespace-available-resources:
    get:
      tags:
      - namespaces
      operationId: forge_namespace_available_resources
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceCatalogItem'
...
  /api/v1/pipelines/{pipeline_id}/badge.svg:
    get:
      tags:
      - pipelines
      operationId: pipeline_badge
      parameters:
      - name: pipeline_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: SVG badge
...
components:
  schemas:
...
    CatalogPage:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CatalogRepository'
...
    LoginRequest:
      type: object
      required:
      - username
      - password
      properties:
        password:
          type: string
        username:
          type: string
    LogoutRequest:
      type: object
      required:
      - refresh_token
      properties:
        refresh_token:
          type: string
          description: The refresh token issued by /auth/login or /auth/refresh.
    LogoutResponse:
      type: object
      required:
      - revoked
      properties:
        revoked:
          type: boolean
...
    RefreshRequest:
      type: object
      required:
      - refresh_token
      properties:
        refresh_token:
          type: string
          description: The previously issued refresh token.
...
    TokenPair:
      type: object
      required:
      - access_token
      - expires_at
      - refresh_token
      properties:
        access_token:
          type: string
        expires_at:
          type: integer
          format: int64
          description: unix seconds
        refresh_token:
          type: string
...